refactor: relocate shared packages into internal/
Move config and prometheus out of cmd/argo-watcher/, and updater and
client out of pkg/, into internal/. This removes the dependency
inversion where internal/ and pkg/ packages imported upward into a
cmd/ binary directory: config was a de-facto shared kernel misfiled
under cmd/, and neither pkg/ package was ever externally importable
(pkg/client imports internal/*, pkg/updater has no external consumers).

Pure package rename plus import-path rewrite with unchanged package
clauses; no behavior change. Build config (Taskfile, codecov,
pre-commit, workflows, trufflehog, gitignore) and one docs link are
updated to the new paths. Generated mocks remain gitignored/untracked.
