+++ .github/workflows/security.yml
@@ -34,7 +34,7 @@ jobs:
         run: task install-deps
 
       - name: Generate mocks
-        # The mock packages (cmd/argo-watcher/mock, pkg/updater/mock) are
+        # The mock packages (cmd/argo-watcher/mock, internal/updater/mock) are
         # gitignored and generated by mockgen. gosec and govulncheck load the
         # full package graph (including _test.go files that import the mocks),
         # so they must exist before the scanners run.
