+++ cmd/argo-watcher/main.go
@@ -5,7 +5,8 @@ import (
 	"log"
 
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/shini4i/argo-watcher/cmd/argo-watcher/config"
+
+	"github.com/shini4i/argo-watcher/internal/config"
 	"github.com/shini4i/argo-watcher/internal/migrate"
 	"github.com/shini4i/argo-watcher/internal/server"
 )
