+++ internal/argocd/argo.go
@@ -7,8 +7,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/shini4i/argo-watcher/cmd/argo-watcher/prometheus"
 	"github.com/shini4i/argo-watcher/internal/helpers"
+	"github.com/shini4i/argo-watcher/internal/prometheus"
 	"github.com/shini4i/argo-watcher/internal/state"
 
 	"github.com/shini4i/argo-watcher/internal/models"
