+++ internal/argocd/argo_api_test.go
@@ -13,10 +13,11 @@ import (
 	"testing"
 	"time"
 
-	"github.com/shini4i/argo-watcher/cmd/argo-watcher/config"
-	"github.com/shini4i/argo-watcher/internal/models"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	"github.com/shini4i/argo-watcher/internal/config"
+	"github.com/shini4i/argo-watcher/internal/models"
 )
 
 type roundTripperFunc func(*http.Request) (*http.Response, error)
