+++ internal/client/utility_test.go
