+++ internal/models/argo_git_backoff_test.go
@@ -8,9 +8,10 @@ import (
 
 	"github.com/go-git/go-git/v5"
 	"github.com/go-git/go-git/v5/plumbing/transport/ssh"
-	"github.com/shini4i/argo-watcher/pkg/updater"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	"github.com/shini4i/argo-watcher/internal/updater"
 )
 
 // fakeGitHandler drives runGitUpdateWithRetry without a real remote: PlainOpen
